# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Initiation Service
  description: API for Payment Initiation based on BIAN PaymentOrder
  version: 1.0.0
servers:
  - url: http://localhost:9999
    description: Local server

paths:
  /payment-initiation/payment-orders:
    post:
      summary: Initiate a payment order
      operationId: initiatePaymentOrder
      tags:
        - Payment Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentOrderRequest'
      responses:
        '201':
          description: Payment order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payment-initiation/payment-orders/{paymentOrderId}:
    get:
      summary: Retrieve a payment order
      operationId: retrievePaymentOrder
      tags:
        - Payment Order
      parameters:
        - name: paymentOrderId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the payment order
      responses:
        '200':
          description: Payment order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderDetails'
        '404':
          description: Payment order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payment-initiation/payment-orders/{paymentOrderId}/status:
    get:
      summary: Retrieve payment order status
      operationId: retrievePaymentOrderStatus
      tags:
        - Payment Order
      parameters:
        - name: paymentOrderId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the payment order
      responses:
        '200':
          description: Payment order status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderStatus'
        '404':
          description: Payment order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PaymentOrderRequest:
      type: object
      required:
        - externalId
        - debtorIban
        - creditorIban
        - amount
        - currency
        - requestedExecutionDate
      properties:
        externalId:
          type: string
          description: External identifier for the payment
        debtorIban:
          type: string
          pattern: '^[A-Z]{2,2}[0-9]{2,2}[a-zA-Z0-9]{1,30}$'
        creditorIban:
          type: string
          pattern: '^[A-Z]{2,2}[0-9]{2,2}[a-zA-Z0-9]{1,30}$'
        amount:
          type: number
          format: double
          minimum: 0.01
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: 'EUR'
        remittanceInfo:
          type: string
        requestedExecutionDate:
          type: string
          format: date

    PaymentOrderResponse:
      type: object
      properties:
        paymentOrderId:
          type: string
        status:
          type: string
          enum: [INITIATED, PENDING, PROCESSED, REJECTED]

    PaymentOrderDetails:
      allOf:
        - $ref: '#/components/schemas/PaymentOrderRequest'
        - type: object
          properties:
            paymentOrderId:
              type: string
            status:
              type: string
            creationDateTime:
              type: string
              format: date-time

    PaymentOrderStatus:
      type: object
      properties:
        paymentOrderId:
          type: string
        status:
          type: string
        lastUpdate:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string